dom/base/test/test_xhr_withCredentials.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=814050
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 814050</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=814050">Mozilla Bug 814050</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 814050, but modified now that the spec has changed **/
var xhr = new XMLHttpRequest();
xhr.open("GET", "", false);
xhr.withCredentials = true;
ok(true, "Should not throw on withCredentials sets for sync XHR");

var xhr = new XMLHttpRequest();
xhr.open("GET", "", true);
xhr.withCredentials = true;
ok(true, "Should not throw on withCredentials sets for async XHR");

</script>
</pre>
</body>
</html>
